feat(functions): Created a Basic Internal Logging and Metrics System
This will likely be expanded out later, but this way we can keep track of metrics from the Firebase logs for which endpoint exactly is called (more than we get from just knowing which Express app is called). We also get runtime metrics and exit code metrics when a function run is complete.