feat(functions): Refactor Express Structure for Logging and Auth (#3)
* refactor(functions): Change Module Function Definitions to Support MW

Now, universal middleware functions can run after a route, which is useful for gathering metrics and logging.

* refactor(functions): Created Central Function to Initialize Modules

This allows all modules to be configured with the same middleware and Firebase setup, with significantly less repeated code.

* feat(functions): Created a Basic Internal Logging and Metrics System

This will likely be expanded out later, but this way we can keep track of metrics from the Firebase logs for which endpoint exactly is called (more than we get from just knowing which Express app is called). We also get runtime metrics and exit code metrics when a function run is complete.

* feat(functions): Create a Mock of the Unified Authorization System

This allows the rest of the code being written for the STPM to check for permissions, even before the entire UAS is complete.