feat(functions): Added a UAS function to handle endpoint permissions
This new function allows you to pass in the logic you want to occur for each type of authorized user, without needing to handle the error handling if the user is not authorized or even authenticated.

A mock function for getCompanyBenefits is available as an example of how to configure and use this new library function.