feat: Create Webapp API Client and API Explorer; UAS is now live (#6)
* fix(functions/uas): Switched "accepted" property from binary to ternary

Now, accepted = true, pending = undefined, and rejected = false. This allows for more granular permission control as to how to handle someone who is not yet approved, and is a key improvement to be able to accept hackers, mentors, and volunteers on an application basis for the MRM, VRM, and HRM.

* feat(functions/uas): uasTokenParser gets user auth information

Note: this is just the basic token validation. It does not check for revoked tokens.

* feat(webapp): Created a custom API Client that includes auth info

* fix(webapp): Fix the Functions Emulator endpoint in Firebase Config

The emulator runs over HTTP, not HTTPS...

* fix(functions/uas): Fix Enum Typo, and permissionSwitch Timeout

Hint: it's helpful if you call the "execute" function...

* chore(webapp/devTools): Create a new FirebaseApp for Dev Tools

This allows things like Auth to be handled separately from the rest of the UI.

* fix(webapp/ApiClient): Fixed Method Passing, Errors, Timeout, and more

Also now allows the FirebaseApp to be configurable, so you can have one client for the main app, and one client for testing or dev tools.

* feat(webapp): Created a API Explorer on localhost

This allows users to test the API calls with Firebase Auth already built in.

* feat(functions/uas): UAS is now live, and no longer stubbed out.