chore: Partition GitHub Actions Build and Test Workflows
Now the `functions` workflows only trigger when changes are made in the `functions` directory.

Now the `webapp` workflows only trigger when changes are made in the `webapp` directory.

To ensure that Branch Protection works as expected with these restrictions, a `default` workflow was created for changes that do not fit in the above categories.

This change also runs the workflows on push to main to get the status for the badges on the README.